use int64, added masks to parameters

+++ spaceship/include/amplifier.h
@@ -7,16 +7,16 @@
 namespace AdventOfCode {
 
 struct AmplifierChain {
-  static int maxThrusterSignal(Program const &program);
+  static int64_t maxThrusterSignal(Program const &program);
 };
 
 class FeedbackLoop {
 public:
   FeedbackLoop(Program const &program);
 
-  int maxOutputSignal ();
+  int64_t maxOutputSignal ();
 private:
-  int outputSignal (Input const& phaseSetting);
+  int64_t outputSignal (Input const& phaseSetting);
 
   std::vector <Computer> amplifiers_;
   Program program_;
