+++ spaceship/src/amplifier.cpp
@@ -5,8 +5,8 @@
 using namespace AdventOfCode;
 
 namespace {
-int getSignal(Input phaseSetting, Program const &program) {
-  int signal = 0;
+int64_t getSignal(Input phaseSetting, Program const &program) {
+  int64_t signal = 0;
   for (int i = 0; i < 5; ++i) {
     Computer amplifier(program);
     amplifier.writeInput({phaseSetting[i]});
@@ -19,12 +19,12 @@ int getSignal(Input phaseSetting, Program const &program) {
 }
 } // namespace
 
-int AmplifierChain::maxThrusterSignal(Program const &program) {
+int64_t AmplifierChain::maxThrusterSignal(Program const &program) {
   Input phaseSetting{0, 1, 2, 3, 4};
 
-  int maxThrusterSignal = 0;
+  int64_t maxThrusterSignal = 0;
   do {
-    int signal = getSignal(phaseSetting, program);
+    int64_t signal = getSignal(phaseSetting, program);
     if (signal > maxThrusterSignal)
       maxThrusterSignal = signal;
   } while (
@@ -35,12 +35,12 @@ int AmplifierChain::maxThrusterSignal(Program const &program) {
 
 FeedbackLoop::FeedbackLoop(Program const &program) : program_{program} {}
 
-int FeedbackLoop::maxOutputSignal() {
+int64_t FeedbackLoop::maxOutputSignal() {
   Input phaseSetting{5, 6, 7, 8, 9};
 
-  int maxThrusterSignal = 0;
+  int64_t maxThrusterSignal = 0;
   do {
-    int signal = outputSignal(phaseSetting);
+    int64_t signal = outputSignal(phaseSetting);
     if (signal > maxThrusterSignal)
       maxThrusterSignal = signal;
   } while (
@@ -49,7 +49,7 @@ int FeedbackLoop::maxOutputSignal() {
   return maxThrusterSignal;
 }
 
-int FeedbackLoop::outputSignal(Input const &phaseSetting) {
+int64_t FeedbackLoop::outputSignal(Input const &phaseSetting) {
   amplifiers_.clear();
   for (int i = 0; i < 5; ++i) {
     Computer computer{program_};
@@ -58,15 +58,15 @@ int FeedbackLoop::outputSignal(Input const &phaseSetting) {
   }
 
   amplifiers_[0].writeInput({0});
-  int current = 0;
-  int lastOutput = 0;
+  int64_t current = 0;
+  int64_t lastOutput = 0;
   Intcode lastCode = Intcode::Input;
   while (!(current == 0 && lastCode == Intcode::Halt)) {
     do {
       lastCode = amplifiers_[current].runInstruction();
     } while (lastCode != Intcode::Output && lastCode != Intcode::Halt);
 
-    int next = (current + 1) % 5;
+    int64_t next = (current + 1) % 5;
     if (lastCode == Intcode::Output) {
       lastOutput = amplifiers_[current].readOutput();
       amplifiers_[next].writeInput({lastOutput});
